backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.14.0
google-adk==1.2.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-cloud-tasks==2.16.0
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.19
httpx==0.28.1
Pillow==11.1.0
numpy==2.2.2
